app/requirements.txt
uvicorn~=0.30.1
fastapi~=0.111.0
PyPDF2
python-docx
pandas
streamlit~=1.32.0
python-dotenv~=1.0.1
requests~=2.31.0
openpyxl~=3.1.2

